--- a/.parcelrc
+++ b/.parcelrc
@@ -1,0 +1,7 @@
+{
+ 	"extends": ["@parcel/config-default"],
+ 	"transformers": {
+ 		"*.png": ["@parcel/transformer-raw"],
+ 		"manifest.json": ["@parcel/transformer-raw"]
+ 	}
+ }