--- a/package.json
+++ b/package.json
@@ -3,8 +3,8 @@
 	"scripts": {
 		"ava": "rollup test/*.ts --plugin typescript --dir .built/test --chunkFileNames _.mjs --entryFileNames [name].mjs && ava",
 		"build": "run-p build:* --continue-on-error",
+		"build:parcel": "parcel build source/refined-github.ts source/background.ts source/options.html source/resolve-conflicts.ts source/icon.png source/manifest.json --no-cache --dist-dir distribution",
 		"build:typescript": "tsc --noEmit",
-		"build:webpack": "webpack --mode=production",
 		"fix": "run-p 'lint:* -- --fix'",
 		"lint": "run-p lint:* --continue-on-error",
 		"lint:css": "stylelint 'source/**/*.css'",
@@ -17,9 +17,13 @@
 		"test": "run-p ava lint:* build:*",
 		"version": "dot-json distribution/manifest.json version $VER",
 		"watch": "run-p watch:* --continue-on-error",
-		"watch:typescript": "tsc --noEmit --watch --preserveWatchOutput",
-		"watch:webpack": "webpack --mode=development --watch"
+		"watch:parcel": "parcel watch source/refined-github.ts source/background.ts source/options.html source/resolve-conflicts.ts source/icon.png source/manifest.json",
+		"watch:typescript": "tsc --noEmit --watch --preserveWatchOutput"
 	},
+	"browserslist": [
+		"last 1 Chrome versions",
+		"last 1 Firefox versions"
+	],
 	"stylelint": {
 		"extends": "stylelint-config-xo",
 		"rules": {
@@ -76,31 +80,22 @@
 		"@sindresorhus/tsconfig": "^0.7.0",
 		"@types/chrome": "0.0.122",
 		"@types/codemirror": "0.0.97",
-		"@types/copy-webpack-plugin": "^6.0.0",
 		"@types/firefox-webext-browser": "^78.0.1",
 		"@types/jsdom": "^16.2.3",
-		"@types/mini-css-extract-plugin": "^0.9.1",
 		"@types/react": "^16.9.46",
-		"@types/terser-webpack-plugin": "^4.1.0",
 		"ava": "^3.12.1",
 		"chrome-webstore-upload-cli": "^1.2.0",
-		"copy-webpack-plugin": "^6.0.3",
-		"css-loader": "^4.2.2",
 		"daily-version": "^2.0.0",
 		"dot-json": "^1.2.0",
 		"eslint-config-xo-react": "^0.23.0",
 		"eslint-plugin-react": "^7.20.6",
 		"eslint-plugin-react-hooks": "^4.1.0",
 		"jsdom": "^16.4.0",
-		"mini-css-extract-plugin": "^0.10.0",
 		"npm-run-all": "^4.1.5",
+		"parcel": "^2.0.0-nightly.416",
 		"rollup": "^2.26.5",
-		"size-plugin": "^2.0.1",
-		"string-replace-loader": "^2.3.0",
 		"stylelint": "^13.6.1",
 		"stylelint-config-xo": "^0.19.0",
-		"terser-webpack-plugin": "^4.1.0",
-		"ts-loader": "^8.0.3",
 		"ts-node": "^9.0.0",
 		"type-fest": "^0.16.0",
 		"typescript": "^4.0.2",
@@ -109,6 +104,15 @@
 		"webpack": "^4.44.1",
 		"webpack-cli": "^3.3.12",
 		"xo": "^0.33.0"
+	},
+	"alias": {
+		"octicon": "@primer/octicons-v2/build/svg",
+		"react": "dom-chef"
+	},
+	"targets": {
+		"default": {
+			"publicUrl": "./"
+		}
 	},
 	"webExt": {
 		"sourceDir": "distribution",