--- a/source/background.ts
+++ b/source/background.ts
@@ -1,3 +1,4 @@
+import 'webextension-polyfill';
 import 'webext-dynamic-content-scripts';
 import cache from 'webext-storage-cache'; // Also needed to regularly clear the cache
 import addDomainPermissionToggle from 'webext-domain-permission-toggle';