--- a/source/features/action-used-by-link.tsx
+++ b/source/features/action-used-by-link.tsx
@@ -1,6 +1,6 @@
 import React from 'dom-chef';
 import select from 'select-dom';
-import SearchIcon from 'octicon/search.svg';
+import {SearchIcon} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
 
 import features from '.';