--- a/source/features/cleanup-repo-filelist-actions.tsx
+++ b/source/features/cleanup-repo-filelist-actions.tsx
@@ -1,7 +1,9 @@
 import React from 'dom-chef';
 import select from 'select-dom';
-import PlusIcon from 'octicon/plus.svg';
-import SearchIcon from 'octicon/search.svg';
+import {
+	PlusIcon,
+	SearchIcon
+} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
 
 import features from '.';