--- a/source/features/conversation-links-on-repo-lists.tsx
+++ b/source/features/conversation-links-on-repo-lists.tsx
@@ -1,8 +1,10 @@
 import React from 'dom-chef';
 import select from 'select-dom';
-import IssueIcon from 'octicon/issue-opened.svg';
+import {
+	IssueOpenedIcon,
+	GitPullRequestIcon
+} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
-import PullRequestIcon from 'octicon/git-pull-request.svg';
 
 import features from '.';
 import observeElement from '../helpers/simplified-element-observer';
@@ -27,13 +29,13 @@ function init(): void {
 				className="muted-link mr-3"
 				href={repositoryLink.href + '/issues?q=is%3Aissue+is%3Aopen'}
 			>
-				<IssueIcon/>
+				<IssueOpenedIcon/>
 			</a>,
 			<a
 				className="muted-link mr-3"
 				href={repositoryLink.href + '/pulls?q=is%3Apr+is%3Aopen'}
 			>
-				<PullRequestIcon/>
+				<GitPullRequestIcon/>
 			</a>
 		);
 	}