--- a/source/features/default-branch-button.tsx
+++ b/source/features/default-branch-button.tsx
@@ -1,7 +1,7 @@
 import React from 'dom-chef';
 import elementReady from 'element-ready';
 import * as pageDetect from 'github-url-detection';
-import ChevronLeftIcon from 'octicon/chevron-left.svg';
+import {ChevronLeftIcon} from '@primer/octicons-react';
 
 import features from '.';
 import GitHubURL from '../github-helpers/github-url';