--- a/source/features/edit-files-faster.css
+++ b/source/features/edit-files-faster.css
@@ -1,7 +1,7 @@
 /* Only shows one icon at a time, depending on hover status */
 .rgh-edit-files-faster:not(:hover) .octicon-pencil,
 .rgh-edit-files-faster:hover > :first-child {
-	display: none;
+	display: none !important;
 }
 
 /* Widen column to avoid bumps in directory-less listings */