--- a/source/features/forked-to.tsx
+++ b/source/features/forked-to.tsx
@@ -5,9 +5,11 @@ import select from 'select-dom';
 import elementReady from 'element-ready';
 import * as pageDetect from 'github-url-detection';
 
-import ForkIcon from 'octicon/repo-forked.svg';
-import CheckIcon from 'octicon/check.svg';
-import LinkExternalIcon from 'octicon/link-external.svg';
+import {
+	RepoForkedIcon,
+	CheckIcon,
+	LinkExternalIcon
+} from '@primer/octicons-react';
 
 import features from '.';
 import fetchDom from '../helpers/fetch-dom';
@@ -86,7 +88,7 @@ async function updateUI(forks: string[]): Promise<void> {
 							title={`Open your fork at ${fork}`}
 						>
 							<span className="select-menu-item-icon rgh-forked-to-icon">
-								{fork === getRepoURL() ? <CheckIcon/> : <ForkIcon/>}
+								{fork === getRepoURL() ? <CheckIcon/> : <RepoForkedIcon/>}
 							</span>
 							{fork}
 						</a>