--- a/source/features/pr-filters.tsx
+++ b/source/features/pr-filters.tsx
@@ -2,7 +2,7 @@ import React from 'dom-chef';
 import cache from 'webext-storage-cache';
 import select from 'select-dom';
 import delegate from 'delegate-it';
-import CheckIcon from 'octicon/check.svg';
+import {CheckIcon} from '@primer/octicons-react';
 import elementReady from 'element-ready';
 import * as pageDetect from 'github-url-detection';