--- a/source/features/release-download-count.tsx
+++ b/source/features/release-download-count.tsx
@@ -1,7 +1,7 @@
 import './release-download-count.css';
 import React from 'dom-chef';
 import select from 'select-dom';
-import DownloadIcon from 'octicon/download.svg';
+import {DownloadIcon} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
 
 import features from '.';