--- a/source/features/repo-age.tsx
+++ b/source/features/repo-age.tsx
@@ -2,7 +2,7 @@ import twas from 'twas';
 import cache from 'webext-storage-cache';
 import React from 'dom-chef';
 import select from 'select-dom';
-import RepoIcon from 'octicon/repo.svg';
+import {RepoIcon} from '@primer/octicons-react';
 import elementReady from 'element-ready';
 import * as pageDetect from 'github-url-detection';