--- a/source/features/show-associated-branch-prs-on-fork.tsx
+++ b/source/features/show-associated-branch-prs-on-fork.tsx
@@ -2,9 +2,11 @@ import React from 'dom-chef';
 import cache from 'webext-storage-cache';
 import onetime from 'onetime';
 import {observe} from 'selector-observer';
-import MergeIcon from 'octicon/git-merge.svg';
+import {
+	GitMergeIcon,
+	GitPullRequestIcon
+} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
-import PullRequestIcon from 'octicon/git-pull-request.svg';
 
 import features from '.';
 import * as api from '../github-helpers/api';
@@ -78,7 +80,7 @@ async function init(): Promise<void> {
 			const branchName = branchCompareLink.closest('[branch]')!.getAttribute('branch')!;
 			const prInfo = associatedPullRequests[branchName];
 			if (prInfo) {
-				const StateIcon = prInfo.state === 'Merged' ? MergeIcon : PullRequestIcon;
+				const StateIcon = prInfo.state === 'Merged' ? GitMergeIcon : GitPullRequestIcon;
 
 				branchCompareLink.replaceWith(
 					<div className="d-inline-block text-right ml-3">