--- a/source/features/wait-for-build.tsx
+++ b/source/features/wait-for-build.tsx
@@ -3,7 +3,7 @@ import React from 'dom-chef';
 import select from 'select-dom';
 import onetime from 'onetime';
 import delegate from 'delegate-it';
-import InfoIcon from 'octicon/info.svg';
+import {InfoIcon} from '@primer/octicons-react';
 import * as pageDetect from 'github-url-detection';
 
 import features from '.';