--- a/source/options.html
+++ b/source/options.html
@@ -67,5 +67,4 @@
 	</div>
 
 </form>
-<script src="browser-polyfill.min.js"></script>
-<script src="options.js"></script>
+<script src="options.tsx"></script>