--- a/source/options.tsx
+++ b/source/options.tsx
@@ -1,4 +1,5 @@
-import 'webext-base-css/webext-base.css';
+import 'webextension-polyfill';
+import 'webext-base-css';
 import './options.css';
 import React from 'dom-chef';
 import cache from 'webext-storage-cache';