… octicon-svg-loader.ts