Follows #2983 and #3618 

Based on #3227 (most of the _changed files_ can be ignored)

Parcel is 3 times slower than Webpack.
Fully-minified filesizes are comparable.

Given the speed drop, I'm not sure if it's worth spending time on the remaining issues.

## Before

```js
❯ nr build:webpack
Time: 4544ms
                   Asset      Size
       refined-github.js   291 KiB
      refined-github.css  40.7 KiB
           background.js  33.7 KiB
            options.html  2.19 KiB
             options.css   3.5 KiB
              options.js  99.2 KiB
    resolve-conflicts.js  2.43 KiB
                icon.png  1.89 KiB
           manifest.json   1.1 KiB
 browser-polyfill.min.js  9.87 KiB
```

## After

```js
❯ nr build:parcel
✨ Built in 15.78s

distribution/refined-github.js         266.81 KB    6.85s
distribution/refined-github.css          25.1 KB    4.65s
distribution/background.js              34.84 KB    6.57s
distribution/options.html                   2 KB    6.85s
distribution/options.4fc4ce54.css        2.41 KB    2.70s
distribution/options.9f319f47.js        55.07 KB    2.86s
distribution/resolve-conflicts.js        1.53 KB    2.86s
distribution/icon.1f68ed73.png           1.89 KB     20ms
distribution/manifest.0c1d4e4f.json      1.04 KB     37ms
```

## Next steps

- [ ] Port the [dynamic variable replacer](https://github.com/sindresorhus/refined-github/blob/0e50ebcf0632680d2e654a719214280a5aa3bc29/webpack.config.ts#L94-L106) to Parcel or Babel
- [ ] Filenames of PNG and JSON files need to be preserved (they're currently hashed)
- [ ] Possibly use the Parcel-native TypeScript type-checker (called Validator)
- [ ] Find ways to speed up?