@fregante 2020-10-05
Perhaps we should just wait for ES Modules support in Firefox (https://bugzilla.mozilla.org/show_bug.cgi?id=1451545) so we can just use Snowpack.