@notlmn 2020-10-05
Options we have right now:

- I hate `webpack` as much as I love it (simple to use, but doesn't have first-class HTML entrypoints)
- `parcel` being parcel is bloated as always and hence the 1/3rd performance compared to webpack
- `rollup` is awesome, but requires us to write custom config the size of a linux kernel
- `snowpack` sounds promising, would be worth a try if it lives up to its reputation

The reason I would like to keep sticking to webpack (for now) is the ease of use, __it just works__ (most of the time though).

Also relevant to the discussion at https://github.com/notlmn/browser-extension-template/issues/34

---------

> ## Next steps
> - Port the dynamic variable replacer to Parcel or Babel
> - Filenames of PNG and JSON files need to be preserved (they're currently hashed)
> - Possibly use the Parcel-native TypeScript type-checker (called Validator)
> - Find ways to speed up?

These do require writing a config file anyway (I assume) which needs to be maintained, so we're right back to the question if we want to keep maintaining the `webpack` config we have.