@fregante 2020-10-05
> if we want to keep maintaining the `webpack` config we have.

That's what I thought. For Refined GitHub, changing builder means porting/rewriting the feature list injection so it takes effort that the change of builder is supposed to prevent.

I had also tried Rollup (https://github.com/sindresorhus/refined-github/pull/3019) but, yeah, it needed just as much configuration and is as slow as Parcel (https://github.com/rollup/rollup/issues/3506). The only advantage is better code output.

https://github.com/sindresorhus/refined-github/pull/3227 itself will drop the need for an SVG loader, so the config will be reduced slightly.

---

However, if you remove SVG import (which is supported via SVGR in Parcel 2 and [supposedly supported](https://github.com/sindresorhus/refined-github/pull/3618/files) in Parcel 2) and feature list injection (which is a very custom solution anyway), Parcel 1 would be good enough for most extensions thanks to [parcel-plugin-web-extension](https://github.com/kevincharm/parcel-plugin-web-extension), as described in https://github.com/notlmn/browser-extension-template/issues/34